% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{kuromoji}
\alias{kuromoji}
\title{Japanese Part-of-Speech and Morphological Analyze using Kuromoji}
\usage{
\S4method{kuromoji}{jaNLP}(str = "", set.mode = NULL, jar.path = NULL)
}
\arguments{
\item{\code{str}}{input single character vector for Kuromoji analysis}

\item{\code{set.mode}}{set Kuromoji's segmentation modes (default: "Normal"). See \url{http://www.atilika.org}[Kuromoji] "Designed for search"}

\item{\code{jar.path}}{kuromoji.jar file path (default: "./libs/kuromoji-0.7.7.jar")}
}
\value{
\describe{
 \item{\code{surface}}{表層形}
 \item{\code{feature}}{品詞,品詞細分類1,品詞細分類2,品詞細分類3,活用形,活用型,原形,読み,発音}
 \item{\code{is_know}}{"Returns true if this token is known word"}
 \item{\code{is_unk}}{"Returns true if this token is unknown word"}
 \item{\code{is_user}}{"Returns true if this token is defined in user dictionary"}
 \item{\code{mode}}{segmentation modes}
 }
}
\description{
The \code{kuromoji} method of \code{jaNLP} class for Japanese PoS tagging and morphological analyze
}
\details{
"Kuromoji is an open source Japanese morphological analyzer written in Java." \cr \url{http://www.atilika.org}
}
\examples{
library(RjaNLP)
example <- JaNLP$new(
 text = "日本経済新聞でモバゲーの記事を読んだ",
 kuromoji_jar = "./share/kuromoji-0.7.7.jar"
)
normal.res <- example$kuromoji()
search.res <- example$kuromoji(set.mode = "SEARCH")
extend.res <- example$kuromoji(set.mode = "EXTENDED")
}

