% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{mecab}
\alias{mecab}
\title{Japanese Part-of-Speech and Morphological Analyze using MeCab}
\usage{
\S4method{mecab}{jaNLP}(str = "")
}
\arguments{
\item{\code{str}}{input single character vector for MeCab analysis}
}
\value{
\describe{
 \item{\code{surface}}{表層形}
 \item{\code{posid}}{pos-id \url{http://mecab.googlecode.com/svn/trunk/mecab/doc/posid.html}}
 \item{\code{feature}}{品詞,品詞細分類1,品詞細分類2,品詞細分類3,活用形,活用型,原形,読み,発音}
 \item{\code{char_type}}{char.def}
 \item{\code{stat}}{MECAB_NOR=0, MECAB_UNK=1, MECAB_BOS=2, MECAB_EOS=3}
 \item{\code{cost}}{cumulative cost}
 \item{\code{wcost}}{word occur cost}
 \item{\code{alpha}}{foward log prob}
 \item{\code{beta}}{backward log prob}
 \item{\code{prob}}{marginal probability}
 }
}
\description{
The \code{mecab} method of \code{jaNLP} class for Japanese PoS tagging and morphological analyze
}
\details{
MeCab is Part-of-Speech and Morphological Analyzer created by Taku Kudo \cr \url{https://code.google.com/p/mecab/} \cr \cr The \code{mecab} method using MeCab-Option "-l 2" \cr See \url{http://mecab.googlecode.com/svn/trunk/mecab/doc/index.html}
}
\examples{
library(RjaNLP)
example <- JaNLP$new(text = "すもももももももものうち")
res.mecab <- example$mecab()
example <- JaNLP$new(text = "")
example$mecab()
}

