% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{RjaNLP-package}
\alias{RjaNLP-package}
\title{Japanese NLP package for R}
\description{
RjaNLP is Natural Language Processing package for Japanese documents
}
\details{
This package is NLP tool(MeCab, CaboCha and so on) wrapper object \cr and include text data process method
}
\author{
Author: Yoshiaki AMANO. \cr
Maintainer: Yoshiaki AMANO. \cr
Twitter: @yamano357
}
\seealso{
\code{\link[openNLP:Parse_Annotator]{openNLP::Parse_Annotator}} \cr
\code{\link[NLP:annotators]{NLP::annotators}} \cr
\code{\link[tm:tokenizer]{tm::tokenizer}}
}
\keyword{documentation}

